refactor(library): drop render-time route interception, keep navigation-time rewrite
The beforeNavigate hook in library/+layout.svelte is the single
list->detail mechanism: it reads window.location at navigation time, so
text searches typed after render are always captured. The resolvePath
interception was a redundant render-time layer that also rewrote
navigation it did not own; resolvePath returns to a pass-through.

getSafeLibraryReturnTo moves into libraryReturnNavigation.ts next to the
rewrite logic it validates, with tests ported from routing.test.ts.
