+++ src/lib/utils/libraryReturnNavigation.test.ts
@@ -2,7 +2,8 @@ import { describe, expect, it } from 'vitest';
 
 import {
 	getLibraryDetailWithReturnTo,
-	getLibraryNavigationContext
+	getLibraryNavigationContext,
+	getSafeLibraryReturnTo
 } from './libraryReturnNavigation';
 
 describe('library return navigation', () => {
@@ -78,5 +79,39 @@ describe('library return navigation', () => {
 				'https://example.com/library/tv/series-id'
 			)
 		).toBeNull();
+	});
+});
+
+describe('getSafeLibraryReturnTo', () => {
+	it('restores the exact filtered movie list URL', () => {
+		expect(
+			getSafeLibraryReturnTo(
+				'/library/movies?monitored=unmonitored&hdrFormat=dolby-vision&sort=year-desc&q=matrix',
+				'/library/movies'
+			)
+		).toBe('/library/movies?monitored=unmonitored&hdrFormat=dolby-vision&sort=year-desc&q=matrix');
+	});
+
+	it('restores the exact filtered TV list URL', () => {
+		expect(
+			getSafeLibraryReturnTo('/library/tv?status=ended&resolution=1080p&sort=size-desc&q=voyager', '/library/tv')
+		).toBe('/library/tv?status=ended&resolution=1080p&sort=size-desc&q=voyager');
+	});
+
+	it('does not use a movie return path for the TV list', () => {
+		expect(getSafeLibraryReturnTo('/library/movies?fileStatus=missingFile', '/library/tv')).toBeNull();
+	});
+
+	it('rejects external and protocol-relative return paths', () => {
+		expect(
+			getSafeLibraryReturnTo('https://example.com/library/movies', '/library/movies')
+		).toBeNull();
+		expect(getSafeLibraryReturnTo('//example.com/library/movies', '/library/movies')).toBeNull();
+	});
+
+	it('rejects missing and malformed values', () => {
+		expect(getSafeLibraryReturnTo(null, '/library/movies')).toBeNull();
+		expect(getSafeLibraryReturnTo('', '/library/movies')).toBeNull();
+		expect(getSafeLibraryReturnTo('not-a-path', '/library/movies')).toBeNull();
 	});
 });
